+++ src/main/java/kitchenpos/ordertable/domain/OrderTableRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.dao;
+package kitchenpos.ordertable.domain;
 
-import kitchenpos.domain.OrderTable;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.repository.Repository;
