+++ src/main/java/kitchenpos/ui/OrderRestController.java
@@ -1,7 +1,7 @@
 package kitchenpos.ui;
 
-import kitchenpos.application.OrderService;
-import kitchenpos.dto.OrderDto;
+import kitchenpos.order.service.OrderService;
+import kitchenpos.order.service.OrderDto;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
