+++ src/test/java/kitchenpos/OrderIntegrationTest.java
@@ -5,7 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 import io.restassured.RestAssured;
 import java.util.List;
-import kitchenpos.dto.OrderDto;
+import kitchenpos.order.service.OrderDto;
 import kitchenpos.fixture.MenuFixture;
 import kitchenpos.fixture.MenuGroupFixture;
 import kitchenpos.fixture.OrderFixture;
