+++ src/test/java/kitchenpos/OrderTableIntegrationTest.java
@@ -5,10 +5,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.List;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.dto.OrderDto;
-import kitchenpos.dto.OrderTableDto;
-import kitchenpos.dto.TableGroupDto;
+import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.order.service.OrderDto;
+import kitchenpos.ordertable.service.OrderTableDto;
+import kitchenpos.ordertable.service.TableGroupDto;
 import kitchenpos.fixture.OrderFixture;
 import kitchenpos.fixture.OrderTableFixture;
 import kitchenpos.fixture.TableGroupFixture;
