+++ src/test/java/kitchenpos/TableGroupIntegrationTest.java
@@ -6,7 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.List;
-import kitchenpos.dto.TableGroupDto;
+import kitchenpos.ordertable.service.TableGroupDto;
 import kitchenpos.fixture.MenuFixture;
 import kitchenpos.fixture.MenuGroupFixture;
 import kitchenpos.fixture.OrderTableFixture;
