+++ src/test/java/kitchenpos/fixture/TableGroupFixture.java
@@ -5,10 +5,10 @@ import static java.util.stream.Collectors.toList;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.function.Consumer;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
-import kitchenpos.dto.OrderTableDto;
-import kitchenpos.dto.TableGroupDto;
+import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.ordertable.domain.TableGroup;
+import kitchenpos.ordertable.service.OrderTableDto;
+import kitchenpos.ordertable.service.TableGroupDto;
 
 public enum TableGroupFixture {
