+++ src/test/java/kitchenpos/steps/IntegrationTestSteps.java
@@ -7,10 +7,10 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import kitchenpos.menu.service.MenuDto;
 import kitchenpos.menugroup.service.MenuGroupDto;
-import kitchenpos.dto.OrderDto;
-import kitchenpos.dto.OrderTableDto;
+import kitchenpos.order.service.OrderDto;
+import kitchenpos.ordertable.service.OrderTableDto;
 import kitchenpos.product.service.ProductDto;
-import kitchenpos.dto.TableGroupDto;
+import kitchenpos.ordertable.service.TableGroupDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
